feat: split Settings.tsx into per-section components
Decompose the monolithic web/src/pages/Settings.tsx (~2200 lines, 60+
useState hooks) into per-section components under web/src/pages/settings/,
one per CollapsibleSection: Profile, Training, Notifications, Security,
Integrations, Pokemon, AIAutomation, KioskTokens. Shared types and pace/
time helpers move to settings/types.ts.

Each section owns its own state, effects, and handlers. Settings.tsx is
reduced to a ~180-line orchestrator handling the preferences fetch,
save-toast plumbing, isChild/admin/feature gating, and section
composition. Section order, headings, collapse behavior, and gating are
unchanged; no behavior, copy, or API changes. Adds a gating smoke test.

Bead: Hytte-6p59i
