+++ web/src/pages/settings/KioskTokensSection.tsx
@@ -1,0 +1,7 @@
+import TokenManager from '../../components/kiosk/TokenManager'
+
+function KioskTokensSection() {
+  return <TokenManager />
+}
+
+export default KioskTokensSection
